First Project Upload
My first SQL database is a simple Magic: The Gathering database, featuring tables with information on mana colors, color combinations, all released sets, and a basic deck that references these tables. Included alongside the database is a document where I workshopped all the code necessary for its construction. Additionally, there is an Excel sheet that mirrors the format and structure of the database, initially used as a blueprint and temporary storage while collecting the necessary data.